Fix missing const qualifier on comparator (MSVC)
A const qualifier was missing on
SPIRVDecorateGeneric::Comparator::operator(), which resulted in a
compile-time error on MSVC, as const objects cannot invoke non-const
member functions.  Since both arguments to this function are const, and
this function has no side-effects, this function can be safely made
const, and therefore invokable in const expressions.

This patch enables SPIRV-LLVM to build using Visual Studio 14.0 (MSVC
19)