Changing placement of lowered constant expressions
According to SPIRV specification (p2.4):
"All OpVariable instructions in a function must be in the first block in the
function. These instructions, together with any immediately preceding OpLine
instructions, must be the first instructions in that block."
Therefore, lowered constant expressions must not be inserted at the beginig of
the basic block.
Insert it before the current instruction if the current basic block is the
first one in the function. Otherwise insert it at the end of the first basic
block in the function.
This partially fixes #50