(1/10) Build warnings clean up
gcc 5.4 was warning that on the _SPIRV_OP variadic macro invokes the
list argument should contain at least one element, in compliance
with C++11.

Given that the list was completely ignored in the _SPIRV_OP macro we
provide one extra argument with the "dummy" value for the cases in
which the list was previously empty. The dummy argument is ignored
but provides the element needed to build the code cleanly.