(6/10) Build warnings clean up
Changed "const static char *" for "const char* const" for the purpose
of making explicit that neither the pointer to the strings or the
characters of the string can be modified, and the declarations in
the SPIRVInternal.h are indeed constants.

Making clear to gcc that we are defining constants avoids a warning
when a .cpp file includes SPIRVInternal.h but does not use some of
the strings.