(10/10) Build warning workaround
gcc 5.4 is unable to track down that the RPCArgs variable is used.
In spite the variable is actually used when extracting elements
with the std::get template, the compiler thinks it is unused and
produces a very verbose and useless warning (20 lines).

This patch is a workaround for the warning, and together with the other
9 patches in this batch gives us a warning free builds with gcc 5.4.