Fixed a bug introduced by a previous commit
The constant strings in SPIRVInternal.h were changed from the "const static
char []" type to "const char* const" for removing a warning.

The change was correct but I missed that there was a usage of the sizeof operand
for determining the length of one of the constant strings. Because the new type
is a pointer the sizeof operator always returned 8 (or 4 if building for 32
 bits) irrespective of the string length, whereas before it returned the string
length plus one character.

The sizeof operator has been replaced for strlen that works irrespective of
whether the string is defined as a pointer or as an array.

Signed-off-by: Abel Bernabeu <abel.bernabeu@gmail.com>