Ported the CreatePipeStograge test to LLVM 3.8.1
The CreatePipeStorage.ll test is one of the tests that were failing
on the 3.8.1 branch. This patch fixes the issue.

An analysis of the issue showed that the syntax for the "load" IR
instruction has changed sligthly from LLVM 3.6 to LLVM 3.8.1, being
that the reason why the LLVM assembler output did not match the
expected output.

Signed-off-by: Abel Bernabeu <abel.bernabeu@gmail.com>