Khronos/spirv 3.8 (#216)
* (1/10) Build warnings clean up

gcc 5.4 was warning that on the _SPIRV_OP variadic macro invokes the
list argument should contain at least one element, in compliance
with C++11.

Given that the list was completely ignored in the _SPIRV_OP macro we
provide one extra argument with the "dummy" value for the cases in
which the list was previously empty. The dummy argument is ignored
but provides the element needed to build the code cleanly.

* (2/10) Build warnings clean up

gcc 5.4 warned that the Kind variable was not used, this patch fixes
the issue.

* (3/10) Build warnings clean up

gcc 5.4 warned that there was an unneeded semicolon. This patch fixes it.

* (4/10) Build warnings clean up

gcc 5.4 warned that the const qualifier on the return value from
getSupportedVersion was redundant, as it was a value and not a reference
what it was being returned.

* (5/10) Build warnings clean up

gcc 5.4 suggested to add braces around the outer if statement body for
making more explicit to a human inspecting the code that there are two
nested if statements and the "else" part belongs to the inner "if".

* (6/10) Build warnings clean up

Changed "const static char *" for "const char* const" for the purpose
of making explicit that neither the pointer to the strings or the
characters of the string can be modified, and the declarations in
the SPIRVInternal.h are indeed constants.

Making clear to gcc that we are defining constants avoids a warning
when a .cpp file includes SPIRVInternal.h but does not use some of
the strings.

* (7/10) Build warnings clean up

The "Store" variable was defined but not used.

* (8/10) Build warnings clean up

getMDString was defined but not used.

* (9/10) Build warnings clean up

The assert statement was redundantly checking that an unsigned value
was greater or equal than zero, which is trivially true and produced
a warning with gcc 5.4.

* (10/10) Build warning workaround

gcc 5.4 is unable to track down that the RPCArgs variable is used.
In spite the variable is actually used when extracting elements
with the std::get template, the compiler thinks it is unused and
produces a very verbose and useless warning (20 lines).

This patch is a workaround for the warning, and together with the other
9 patches in this batch gives us a warning free builds with gcc 5.4.

* Fixed a bug introduced with a previous commit

The constant strings in SPIRVInternal.h were changed from the "const static
char []" type to "const char* const" for removing a warning.

The change was correct but I missed that there was a usage of the sizeof operand
for determining the length of one of the constant strings. Because the new type
is a pointer the sizeof operator always returned 8 (or 4 if building for 32
 bits) irrespective of the string length, whereas before it returned the string
length plus one character.

The sizeof operator has been replaced for strlen that works irrespective of
whether the string is defined as a pointer or as an array.

Signed-off-by: Abel Bernabeu <abel.bernabeu@gmail.com>

* Ported the CreatePipeStograge test to LLVM 3.8.1

The CreatePipeStorage.ll test is one of the tests that were failing
on the 3.8.1 branch. This patch fixes the issue.

An analysis of the issue showed that the syntax for the "load" IR
instruction has changed sligthly from LLVM 3.6 to LLVM 3.8.1, being
that the reason why the LLVM assembler output did not match the
expected output.

Signed-off-by: Abel Bernabeu <abel.bernabeu@gmail.com>