~SPIRVModuleImpl() crashes when delete no-id entries
Use type std::set for "EntryNoId" to avoid duplicate elements.The elements
in std:set are sorted by memory pointer, not by creation order. An entry of
OpLine (A) can be deleted because of deletion of another entry (B) which refers
to A. If A is placed after B, it will try to access invalid A when it frees
"EntryNoId". Therefore, not insert entry of OpLine to "EntryNoId", and it will
be deleted by std::shared_ptr automatically.